+++ Assets/_GameAssets/Scripts/Collectibles/EggCollectible.cs
@@ -1,0 +1,12 @@
+using UnityEngine;
+
+public class EggCollectible : MonoBehaviour
+{
+    [SerializeField] private GameManager _gameManager;
+
+    public void Collect()
+    {
+        GameManager.Instance.IncreaseTotalEggCount();
+        Destroy(gameObject);
+    }
+}